Renamed shared example files.
 * removes the last rspec warnings

 * renamed shared_condition_specs as well because it could be confusing.
