+++ spec/shared_examples/condition_responder_timing_shared.rb
