+++ spec/shared_examples/scope_updated_in_date_range.rb
