+++ spec/shared_examples/shared_conditions.rb
